backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.2
sentence-transformers>=2.6
nomic==3.9.0
faiss-cpu>=1.8
joblib>=1.4
PyPDF2>=3.0
python-docx>=1.1
pdf2image>=1.17
pillow>=10.3
pytesseract>=0.3.10
ollama>=0.3.0
unittest-xml-reporting>=3.2
coverage>=7.6
tree_sitter_languages>=1.10
